+++ package.json
@@ -21,6 +21,7 @@
     "@mui/lab": "^7.0.0-beta.11",
     "@mui/material": "^7.0.1",
     "@mui/styled-engine-sc": "^7.0.1",
+    "@radix-ui/react-alert-dialog": "^1.1.14",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-dialog": "^1.1.14",
     "@radix-ui/react-dropdown-menu": "^2.1.15",
